use int8 for strand

+++ src/genomicranges/utils.py
@@ -41,7 +41,7 @@ def sanitize_strand_vector(
             raise ValueError(
                 "'strand' must only contain values 1 (forward strand), -1 (reverse strand) or 0 (reverse strand)."
             )
-        return strand
+        return strand.astype(np.int8)
 
     if ut.is_list_of_type(strand, str):
         if not set(strand).issubset(["+", "-", "*"]):
@@ -53,7 +53,7 @@ def sanitize_strand_vector(
             raise ValueError(
                 "'strand' must only contain values 1 (forward strand), -1 (reverse strand) or 0 (reverse strand)."
             )
-        return np.array(strand)
+        return np.array(strand, dtype=np.int8)
 
     raise TypeError(
         "'strand' must be either a numpy vector, a list of integers or strings representing strand."
